+++ Binance.Api/Options/Clients/BinanceOptionsRestClient.cs
@@ -24,7 +24,7 @@ internal partial class BinanceOptionsRestClient : IBinanceOptionsRestClient
         MarketMaker = new BinanceOptionsRestClientMarketMaker(this);
     }
 
-    private Task<RestCallResult<T>> RequestAsync<T>(
+    internal Task<RestCallResult<T>> RequestAsync<T>(
         Uri uri, HttpMethod method, CancellationToken cancellationToken, bool signed = false,
         Dictionary<string, object>? queryParameters = null,
         Dictionary<string, object>? bodyParameters = null,
@@ -35,7 +35,7 @@ internal partial class BinanceOptionsRestClient : IBinanceOptionsRestClient
         int requestWeight = 1) where T : class
         => _.RequestAsync<T>(uri, method, cancellationToken, signed, queryParameters, bodyParameters, headerParameters, serialization, deserializer, ignoreRatelimit, requestWeight);
 
-    private Uri GetUrl(string api, string version, string endpoint)
+    internal Uri GetUrl(string api, string version, string endpoint)
     {
         var url = BinanceAddress.Default.OptionsRestApiAddress;
         if (!string.IsNullOrEmpty(api)) url = url.AppendPath($"{api}");