+++ Binance.Api/Options/Responses/BinanceOptionsMarketMakerUnderlyings.cs
@@ -1,0 +1,12 @@
+﻿namespace Binance.Api.Options;
+
+/// <summary>
+/// Binance Options Market Maker Underlyings
+/// </summary>
+public record BinanceOptionsMarketMakerUnderlyings
+{
+    /// <summary>
+    /// Underlyings
+    /// </summary>
+    public List<string> Underlyings { get; set; } = [];
+}